almacenamiento/requirement.txt
psycopg2-binary
pika
flask
redis